database-feeding-service/requirements.txt
flask==2.0.1
werkzeug==2.0.1
psycopg2-binary